Extracting inner trade component TradeRow
